Split the dependencies into soft and hard ones.
This allows to not fetch all the dependencies of this project when wanting to
only use a subset of the possibilities covered by this module. For instance,
this is true if you don't want to use the macauth repoze.who facility and thus
don't care about getting all the dependencies for this.

However, if you want to install all the dependencies, you still can, by using
the pip-compatible `requirements.txt` file.